Conflicts:
	compile.sh

+++ src/java/nxt/BlockImpl.java
@@ -328,6 +328,7 @@ final class BlockImpl implements Block {
 
     boolean verifyGenerationSignature() throws BlockchainProcessor.BlockOutOfOrderException {
 
+        //kushti: comment below to have fast block generation with any balance
         try {
 
             BlockImpl previousBlock = (BlockImpl)Nxt.getBlockchain().getBlock(this.previousBlockId);
@@ -359,15 +360,13 @@ final class BlockImpl implements Block {
 
             BigInteger hit = new BigInteger(1, new byte[] {generationSignatureHash[7], generationSignatureHash[6], generationSignatureHash[5], generationSignatureHash[4], generationSignatureHash[3], generationSignatureHash[2], generationSignatureHash[1], generationSignatureHash[0]});
 
-            return Generator.verifyHit(hit, effectiveBalance, previousBlock, timestamp) || (this.height < Constants.TRANSPARENT_FORGING_BLOCK_5 && badBlocks.contains(this.getId()));
+            return Generator.verifyHit(hit, BigInteger.valueOf(effectiveBalance), previousBlock, timestamp) || (this.height < Constants.TRANSPARENT_FORGING_BLOCK_5 && badBlocks.contains(this.getId()));
 
         } catch (RuntimeException e) {
 
             Logger.logMessage("Error verifying block generation signature", e);
             return false;
-
         }
-
     }
 
     private static final Set<Long> badBlocks = Collections.unmodifiableSet(new HashSet<>(Arrays.asList(
